@@ -4,42 +4,65 @@
     "description": "Library for (de-)serializing data of any complexity; supports XML, JSON, and YAML.",
     "keywords": ["serialization", "deserialization", "json", "jaxb", "xml"],
     "homepage": "http://jmsyst.com/libs/serializer",
-    "license": "Apache2",
+    "license": "MIT",
     "authors": [
         {
             "name": "Johannes M. Schmitt",
             "email": "schmittjoh@gmail.com"
+        },
+        {
+            "name": "Asmir Mustafic",
+            "email": "goetas@gmail.com"
         }
     ],
     "require": {
-        "php": ">=5.3.2",
-        "jms/metadata": "~1.1",
-        "jms/parser-lib": "1.*",
-        "phpcollection/phpcollection": ">=0.1,<0.3-dev",
-        "doctrine/annotations": "1.*"
+        "php": "^7.2",
+        "jms/metadata": "^2.0",
+        "doctrine/annotations": "^1.0",
+        "doctrine/instantiator": "^1.0.3",
+        "hoa/compiler": "^3.17.08.08"
     },
     "suggest": {
-        "symfony/yaml": "Required if you'd like to serialize data to YAML format."
+        "symfony/yaml": "Required if you'd like to use the YAML metadata format.",
+        "doctrine/collections": "Required if you like to use doctrine collection types as ArrayCollection.",
+        "doctrine/cache": "Required if you like to use cache functionality."
     },
-    "minimum-stability": "dev",
     "require-dev": {
-        "twig/twig": ">=1.8,<2.0-dev",
-        "doctrine/orm": ">=2.1,<2.4-dev",
-        "propel/propel1": "~1.7",
-        "symfony/yaml": "2.*",
-        "symfony/translation": ">=2.0,<2.2-dev",
-        "symfony/validator": ">=2.0,<2.2-dev",
-        "symfony/form": ">=2.1,<2.2-dev",
-        "symfony/filesystem": "2.*"
+        "ext-pdo_sqlite": "*",
+        "twig/twig": "~1.34|~2.4",
+        "doctrine/orm": "~2.1",
+        "jackalope/jackalope-doctrine-dbal": "^1.1.5",
+        "doctrine/phpcr-odm": "^1.3|^2.0",
+        "ocramius/proxy-manager": "^1.0|^2.0",
+        "psr/container": "^1.0",
+        "symfony/dependency-injection": "^3.0|^4.0|^5.0",
+        "symfony/yaml": "^3.3|^4.0|^5.0",
+        "symfony/translation": "^3.0|^4.0|^5.0",
+        "symfony/validator": "^3.1.9|^4.0|^5.0",
+        "symfony/form": "^3.0|^4.0|^5.0",
+        "symfony/filesystem": "^3.0|^4.0|^5.0",
+        "symfony/expression-language": "^3.0|^4.0|^5.0",
+        "phpunit/phpunit": "^7.5||^8.0||^9.0",
+        "doctrine/coding-standard": "^5.0"
+    },
+    "conflict": {
+        "hoa/core": "*",
+        "hoa/consistency": "<1.17.05.02",
+        "hoa/iterator": "<2.16.03.15"
     },
     "autoload": {
-        "psr-0": {
-            "JMS\\Serializer": "src/"
+        "psr-4": {
+            "JMS\\Serializer\\": "src/"
+        }
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "JMS\\Serializer\\Tests\\": "tests/"
         }
     },
     "extra": {
         "branch-alias": {
-            "dev-master": "0.14-dev"
+            "dev-master": "3.6-dev"
         }
     }
 }
