@@ -1,18 +1,26 @@
 <?php
 
-if ( ! isset($_SERVER['argv'][1], $_SERVER['argv'][2])) {
-    echo 'Usage: php benchmark.php <format> <iterations> [output-file]'.PHP_EOL;
+declare(strict_types=1);
+
+use JMS\Serializer\SerializerBuilder;
+use JMS\Serializer\Tests\Fixtures\Author;
+use JMS\Serializer\Tests\Fixtures\BlogPost;
+use JMS\Serializer\Tests\Fixtures\Comment;
+use JMS\Serializer\Tests\Fixtures\Publisher;
+
+if (!isset($_SERVER['argv'][1], $_SERVER['argv'][2])) {
+    echo 'Usage: php benchmark.php <format> <iterations> [output-file]' . PHP_EOL;
     exit(1);
 }
 
-list(, $format, $iterations) = $_SERVER['argv'];
+[, $format, $iterations] = $_SERVER['argv'];
 
 require_once 'bootstrap.php';
 
-function benchmark(\Closure $f, $times = 10)
+function benchmark(Closure $f, $times = 10)
 {
     $time = microtime(true);
-    for ($i=0; $i<$times; $i++) {
+    for ($i = 0; $i < $times; $i++) {
         $f();
     }
 
@@ -21,8 +29,8 @@ function benchmark(\Closure $f, $times = 10)
 
 function createCollection()
 {
-    $collection = array();
-    for ($i=0; $i<50; $i++) {
+    $collection = [];
+    for ($i = 0; $i < 200; $i++) {
         $collection[] = createObject();
     }
 
@@ -31,32 +39,33 @@ function createCollection()
 
 function createObject()
 {
-    $post = new \JMS\Serializer\Tests\Fixtures\BlogPost('FooooooooooooooooooooooBAR', new \JMS\Serializer\Tests\Fixtures\Author('Foo'), new \DateTime);
-    for ($i=0; $i<10; $i++) {
-        $post->addComment(new \JMS\Serializer\Tests\Fixtures\Comment(new \JMS\Serializer\Tests\Fixtures\Author('foo'), 'foobar'));
+    $p = new Publisher('bar');
+    $post = new BlogPost('FooooooooooooooooooooooBAR', new Author('Foo'), new DateTime(), $p);
+    for ($i = 0; $i < 100; $i++) {
+        $post->addComment(new Comment(new Author('foo'), 'foobar'));
     }
 
     return $post;
 }
 
-$serializer = \JMS\Serializer\SerializerBuilder::create()->build();
+$serializer = SerializerBuilder::create()->build();
 $collection = createCollection();
-$metrics = array();
-$f = function() use ($serializer, $collection, $format) {
+$metrics = [];
+$f = static function () use ($serializer, $collection, $format) {
     $serializer->serialize($collection, $format);
 };
 
 // Load all necessary classes into memory.
 benchmark($f, 1);
 
-printf('Benchmarking collection for format "%s".'.PHP_EOL, $format);
-$metrics['benchmark-collection-'.$format] = benchmark($f, $iterations);
+printf('Benchmarking collection for format "%s".' . PHP_EOL, $format);
+$metrics['benchmark-collection-' . $format] = benchmark($f, $iterations);
 
-$output = json_encode(array('metrics' => $metrics));
+$output = json_encode(['metrics' => $metrics]);
 
 if (isset($_SERVER['argv'][3])) {
     file_put_contents($_SERVER['argv'][3], $output);
-    echo "Done.".PHP_EOL;
+    echo 'Done.' . PHP_EOL;
 } else {
-    echo $output.PHP_EOL;
+    echo $output . PHP_EOL;
 }
