@@ -1,30 +1,16 @@
 <?php
 
-/*
- * Copyright 2011 Johannes M. Schmitt <schmittjoh@gmail.com>
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
+declare(strict_types=1);
 
 use Doctrine\Common\Annotations\AnnotationRegistry;
 
-call_user_func(function() {
-    if ( ! is_file($autoloadFile = __DIR__.'/../vendor/autoload.php')) {
-        throw new \RuntimeException('Did not find vendor/autoload.php. Did you run "composer install --dev"?');
+(static function () {
+    if (!is_file($autoloadFile = __DIR__ . '/../vendor/autoload.php')) {
+        throw new RuntimeException('Did not find vendor/autoload.php. Did you run "composer install --dev"?');
     }
 
     $loader = require $autoloadFile;
     $loader->add('JMS\Serializer\Tests', __DIR__);
 
     AnnotationRegistry::registerLoader('class_exists');
-});
+})();
